Move create-test-data to unit test helper directory
